+++ mcp_ha/src/ha/http.test.ts
@@ -14,6 +14,7 @@ function cfg(partial: Partial<AddonConfig> = {}): AddonConfig {
     entityDenylist: [],
     serviceDenylist: [],
     confirmDomains: [],
+    apiTokens: [],
     supervisorToken: null,
     devHaUrl: null,
     devHaToken: null,